back-end/app/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
pdfplumber==0.11.0
rapidfuzz==3.9.7
scikit-learn==1.5.1
sentence-transformers==3.0.1
transformers==4.41.2
pyyaml==6.0.2
python-dotenv==1.0.1
# Install torch from https://pytorch.org/get-started/locally/ for your OS/Python
